chore: moved some things and removed some things
