+++ src/usage/Constants/index.ts
@@ -1,3 +1,5 @@
+export const DUMMY_PRICING_VERSION_TO_DELETE = 4
+
 export const GRAPH_INFO = {
   rateLimits: [
     {
