--- src/usage/utils.ts
@@ -1,1 +1,0 @@
-export const DUMMY_PRICING_VERSION_TO_DELETE = 4
